debug

+++ src/arm_controller_node.cpp
@@ -272,7 +272,7 @@ private:
           if (j == 9) {
             current_jpos_des.at(9) +=
               std::clamp(
-              (deg_30 * std::sin(4 * Pi * phase)) - current_jpos_des.at(9),
+              (float) (deg_30 * std::sin(4 * Pi * phase)) - current_jpos_des.at(9),
               -max_joint_delta_, max_joint_delta_);
           }
